task16/Telegram_Bot/requirements.txt
aiogram==3.15.0
asyncio==3.4.3
python-dotenv==1.0.1
openai==1.57.4
langchain==0.3.9 
langchain_community==0.3.9
langchain-openai==0.2.11 
faiss-cpu==1.9.0.post1
